tests: dt_binding_check: new test
We currently "rely" on Rob Herring's mail bot to provide sanity checking
on patches that touch device tree bindings:
https://lore.kernel.org/netdev/176134124286.2841725.8990137232361008022.robh@kernel.org

and that is sufficient in "public" settings (patches sent to
netdev@vger.kernel.org and devicetree@vger.kernel.org also copied).

But in "private" settings (i.e. individual developers encouraged to do
sanity checking on their own, for example by running ingest_mdir), we
quickly find that Rob's mail bot is apparently closed source, and it has
happened more than once (at least to me) for me to miss newly introduced
dt_binding_check errors even if I did the due dilligence of running that
test locally. So an automated check would be good to have.

The justification for including it in NIPA is that while device tree
maintainers review binding patches, they get applied to the subsystem
tree (in this case netdev).

In terms of implementation, "make dt_binding_check" does not always
nicely print "error" or "warning" on the lines with issues. Furthermore,
the errors are multi-line. So instead of filtering for error lines, here
we filter out the "normal" lines, which contain things such as "make",
"SCHEMA", "CHKDT", "LINT", "DTEX", "DTC" at the beginning.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>